rename file to match content
