+++ lib/main_prod.dart
@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:intl/date_symbol_data_local.dart';
 import 'package:lang_words/app_config.dart';
 import 'package:lang_words/firebase_options_prod.dart';
-import 'package:lang_words/services/words_local_service.dart';
+import 'package:lang_words/services/object_box_service.dart';
 import 'package:lang_words/widgets/initialize_app_future.dart';
 import 'package:lang_words/widgets/my_app.dart';
